# tacs.api/src/main/java/tp/tacs/api/model/ModoDeJuegoModel.java
package tp.tacs.api.model;

import java.util.Objects;
import io.swagger.annotations.ApiModel;
import com.fasterxml.jackson.annotation.JsonValue;
import org.springframework.validation.annotation.Validated;
import javax.validation.Valid;
import javax.validation.constraints.*;

import com.fasterxml.jackson.annotation.JsonCreator;

/**
 * Modo de juego.
 */
public enum ModoDeJuegoModel {
  RAPIDO("Rapido"),
    NORMAL("Normal"),
    EXTENDIDO("Extendido");

  private String value;

  ModoDeJuegoModel(String value) {
    this.value = value;
  }

  @Override
  @JsonValue
  public String toString() {
    return String.valueOf(value);
  }

  @JsonCreator
  public static ModoDeJuegoModel fromValue(String text) {
    for (ModoDeJuegoModel b : ModoDeJuegoModel.values()) {
      if (String.valueOf(b.value).equals(text)) {
        return b;
      }
    }
    return null;
  }
}