# tacs.api/src/main/resources/swagger/api.yaml
openapi: 3.0.1
info:
  title: TP TACS
  version: 1.0.0
servers:
  - url: http://127.0.0.1:8080/api/v1
...
paths:
...
  /partidas:
    get:
      tags:
        - Partidas
      summary: Permite listar partidas. No incluye los datos de juego asociados a las mismas.
      operationId: listarPartidas
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: fechaInicio
          schema:
            type: string
            format: date
          required: false
        - in: query
          name: fechaFin
          schema:
            type: string
            format: date
          required: false
        - in: query
          name: estado
          schema:
            $ref: '#/components/schemas/EstadoDeJuegoModel'
          required: false
        - in: query
          name: ordenarPor
          description: Campos por los que ordenar separados por coma. Los campos soportados son fecha y estado.
          example: "fecha,estado"
          explode: false
          schema:
            type: string
          required: false
        - in: query
          name: tamanioPagina
          schema:
            type: integer
            format: int64
            default: 10
          required: false
        - in: query
          name: pagina
          schema:
            type: integer
            format: int64
            default: 0
          required: false
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                title: listarPartidasResponse
                properties:
                  partidas:
                    type: array
                    items:
                      $ref: '#/components/schemas/PartidaModel'
        400:
          description: Bad request
        403:
          description: Forbidden
    post:
      tags:
        - Partidas
      summary: Crear una nueva partida
      operationId: crearPartida
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              title: crearPartidaBody
              properties:
                idProvincia:
                  type: integer
                  format: int64
                  description: Id de la provincia donde se juega
                cantidadMunicipios:
                  type: integer
                  format: int64
                  description: Cantidad de municipios que se usan
                idJugadores:
                  description: Jugadores
                  type: array
                  items:
                    type: integer
                    format: int64
                modoDeJuego:
                  $ref: '#/components/schemas/ModoDeJuegoModel'
      responses:
        201:
          description: Created
        400:
          description: Bad Request
        403:
          description: Forbidden
        404:
          description: Not found
  /partidas/{idPartida}:
    get:
      tags:
        - Partidas
      summary: "Retorna la partida con id: idPartida. Incluye los datos del juego."
      operationId: getPartida
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: idPartida
          description: Id de la partida
          schema:
            type: integer
            format: int64
          required: true
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartidaModel'
        400:
          description: Bad Request
        404:
          description: Not found
    patch:
      tags:
        - Partidas
      summary: Actualiza el estado de una partida
      operationId: actualizarEstadoPartida
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: idPartida
          schema:
            type: integer
            format: int64
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartidaModel'
      responses:
        200:
          description: Success
        400:
          description: Bad request
        404:
          description: Not found
  /partidas/{idPartida}/movimientos:
    post:
      tags:
        - Partidas
      summary: Mueve gauchos de un municipio a otro
      operationId: moverGauchos
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: idPartida
          schema:
            type: integer
            format: int64
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              title: moverGauchosBody
              properties:
                idMunicipioOrigen:
                  type: integer
                  format: int64
                  description: Id del municipio de origen
                idMunicipioDestino:
                  type: integer
                  format: int64
                  description: Id del municipio de destino
                cantidad:
                  type: integer
                  format: int64
                  description: Cantidad de gauchos a mover
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                title: moverGauchosResponse
                properties:
                  municipioOrigen:
                    $ref: '#/components/schemas/MunicipioEnJuegoModel'
                  municipioDestino:
                    $ref: '#/components/schemas/MunicipioEnJuegoModel'
        400:
          description: Bad request
        404:
          description: Not found
  /partidas/{idPartida}/municipios/{idMunicipio}:
    patch:
      tags:
        - Partidas
      summary: Actualiza el estado de un municipio en juego
      operationId: actualizarMunicipio
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: idPartida
          schema:
            type: integer
            format: int64
          required: true
        - in: path
          name: idMunicipio
          schema:
            type: integer
            format: int64
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MunicipioEnJuegoModel'
      responses:
        200:
          description: Success
        400:
          description: Bad request
        404:
          description: Not found
  /partidas/{idPartida}/ataques:
    post:
      tags:
        - Partidas
      summary: Ataca a un municipio. Al atacar, se ataca con todos los gauchos del municipio atacante.
      operationId: atacarMunicipio
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: idPartida
          schema:
            type: integer
            format: int64
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              title: atacarMunicipioBody
              properties:
                idMunicipioAtacante:
                  type: integer
                  format: int64
                  description: Id del municipio que ataca
                idMunicipioObjetivo:
                  type: integer
                  format: int64
                  description: Id del municipio atacado
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                title: atacarMunicipioResponse
                properties:
                  municipioAtacante:
                    $ref: '#/components/schemas/MunicipioEnJuegoModel'
                  municipioAtacado:
                    $ref: '#/components/schemas/MunicipioEnJuegoModel'
        400:
          description: Bad request
        404:
          description: Not found
  /partidas/{idPartida}/ataques/simulaciones:
    post:
      tags:
        - Partidas
      summary: Ataca a un municipio
      operationId: simularAtacarMunicipio
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: idPartida
          schema:
            type: integer
            format: int64
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              title: simularAtacarMunicipioBody
              properties:
                idMunicipioAtacante:
                  type: integer
                  format: int64
                  description: Id del municipio que ataca
                idMunicipioObjetivo:
                  type: integer
                  format: int64
                  description: Id del municipio atacado
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                title: simularAtacarMunicipioResponse
                properties:
                  exitoso:
                    type: boolean
                    description: Si el ataque fue exitoso en la simulación
        400:
          description: Bad request
        404:
          description: Not found
components:
  schemas:
    PartidaModel:
      type: object
      description: Una partida
      required:
        - id
        - provincia
        - cantidadMunicipios
        - estado
        - jugadores
        - modoDeJuego
        - fecha
      properties:
        id:
          type: integer
          format: int64
          description: Id de la partida
        provincia:
          $ref: '#/components/schemas/ProvinciaModel'
        cantidadMunicipios:
          type: integer
          format: int64
          description: Cantidad de municipios que se usan
        estado:
          $ref: '#/components/schemas/EstadoDeJuegoModel'
        jugadores:
          type: array
          description: Usuarios que juegan esta partida
          items:
            $ref: '#/components/schemas/UsuarioModel'
        modoDeJuego:
          $ref: '#/components/schemas/ModoDeJuegoModel'
        fecha:
          type: string
          format: date
          description: Fecha de inicio de la partida
        informacionDeJuego:
          $ref: '#/components/schemas/DatosDeJuegoModel'
        idGanador:
          type: string
          description: Id del usuario ganador
    DatosDeJuegoModel:
      type: object
      description: Datos de una partida. Incluye su provincia en juego y el usuario que tiene el proximo turno.
      required:
        - provincia
      properties:
        idUsuarioProximoTurno:
          type: integer
          format: int64
        municipios:
          type: array
          description: Municipios en juego que contiene la provincia.
          items:
            $ref: '#/components/schemas/MunicipioEnJuegoModel'
    EstadoDeJuegoModel:
      type: string
      description: Estado de una partida.
      enum:
        - EnProgreso
        - Terminada
        - Cancelada
    ModoDeJuegoModel:
      type: string
      description: Modo de juego.
      enum:
        - Rapido
        - Normal
        - Extendido
    UsuarioModel:
      type: object
      description: Un usuario.
      required:
        - id
        - nombreDeUsuario
      properties:
        id:
          type: integer
          format: int64
        nombreDeUsuario:
          type: string
    ProvinciaModel:
      type: object
      description: Una provincia.
      required:
        - id
        - nombre
      properties:
        id:
          type: integer
          format: int64
          description: Id de la provincia
        nombre:
          type: string
          description: Nombre de la provincia
    MunicipioEnJuegoModel:
      type: object
      description: Municipio en juego.
      properties:
        id:
          type: integer
          format: int64
          description: Id del municipio.
        nombre:
          type: string
          description: Nombre del municipio.
        ubicacion:
          $ref: '#/components/schemas/CoordenadasModel'
        altura:
          type: integer
          format: int64
          description: Altura en metros del centroide del municipio
        duenio:
          $ref: '#/components/schemas/UsuarioModel'
        gauchos:
          type: integer
          format: int64
          description: Cantidad de gauchos en el municipio
        puntosDeDefensa:
          type: integer
          format: int64
        produccionDeGauchos:
          type: integer
          format: int64
        modo:
          $ref: '#/components/schemas/ModoDeMunicipioModel'
        estaBloqueado:
          type: boolean
          description: Si el municipio está bloqueado.
        urlImagen:
          type: string
          description: URL a una imagen del municipio
    CoordenadasModel:
      type: object
      required:
        - lat
        - lon
      description: Par de coordenadas en el mundo.
      properties:
        lat:
          type: integer
          format: int64
          description: Latitud
        lon:
          type: integer
          format: int64
          description: Longitud
    ModoDeMunicipioModel:
      type: string
      description: Modo de operación de un municipio.
      enum:
        - Defensa
        - Produccion
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT